refactor: 🚚 Consolidate all custom Exceptions for integrations into one file for easier re-use.
