+++ nautobot_ssot/integrations/aci/jobs.py
@@ -5,6 +5,7 @@ from django.urls import reverse
 from nautobot.dcim.models import Controller, Location
 from nautobot.extras.jobs import BooleanVar, Job, ObjectVar
 
+from nautobot_ssot.exceptions import ConfigurationError
 from nautobot_ssot.integrations.aci.diffsync.adapters.aci import AciAdapter
 from nautobot_ssot.integrations.aci.diffsync.adapters.nautobot import NautobotAdapter
 from nautobot_ssot.integrations.aci.diffsync.client import AciApi
@@ -12,10 +13,6 @@ from nautobot_ssot.jobs.base import DataMapping, DataSource
 from nautobot_ssot.utils import get_username_password_https_from_secretsgroup, verify_controller_managed_device_group
 
 name = "Cisco ACI SSoT"  # pylint: disable=invalid-name, abstract-method
-
-
-class ConfigurationError(Exception):
-    """Exception thrown when Job configuration is wrong."""
 
 
 class AciDataSource(DataSource, Job):  # pylint: disable=abstract-method, too-many-instance-attributes
